provide a way to run REPL commands outside the REPL (#173)
(cherry picked from commit 7a8d376e02fe8073153e148269cc74be337d3619)
